+++ src/main/java/com/tickatch/ticketservice/ticket/application/dto/TicketRequest.java
@@ -1,0 +1,10 @@
+package com.tickatch.ticketservice.ticket.application.dto;
+
+import java.util.UUID;
+
+public record TicketRequest(
+    UUID reservationId,
+    long seatId,
+    String grade,
+    String seatNumber,
+    Long price) {}
