+++ src/main/java/com/tickatch/ticketservice/ticket/application/service/TicketService.java
@@ -6,10 +6,10 @@ import com.tickatch.ticketservice.ticket.domain.exception.TicketErrorCode;
 import com.tickatch.ticketservice.ticket.domain.exception.TicketException;
 import com.tickatch.ticketservice.ticket.domain.repository.TicketRepository;
 import com.tickatch.ticketservice.ticket.domain.service.ReservationService;
-import com.tickatch.ticketservice.ticket.presentation.dto.TicketActionResponse;
-import com.tickatch.ticketservice.ticket.presentation.dto.TicketDetailResponse;
-import com.tickatch.ticketservice.ticket.presentation.dto.TicketRequest;
-import com.tickatch.ticketservice.ticket.presentation.dto.TicketResponse;
+import com.tickatch.ticketservice.ticket.application.dto.TicketActionResponse;
+import com.tickatch.ticketservice.ticket.application.dto.TicketDetailResponse;
+import com.tickatch.ticketservice.ticket.application.dto.TicketRequest;
+import com.tickatch.ticketservice.ticket.application.dto.TicketResponse;
 import java.util.Optional;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
